+++ src/parsing.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   import_map.c                                       :+:      :+:    :+:   */
+/*   parsing.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: emilien <emilien@student.42lyon.fr>        +#+  +:+       +#+        */
+/*   By: eguelin <eguelin@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/16 14:40:17 by eguelin           #+#    #+#             */
-/*   Updated: 2023/03/04 19:18:11 by emilien          ###   ########lyon.fr   */
+/*   Updated: 2023/04/13 17:55:43 by eguelin          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -88,8 +88,8 @@ static int	ft_import_coord(int fd, int y, t_data *data)
 	i = (y - 1) * data->x_max;
 	while (x <= data->x_max)
 	{
-		data->map[i].x = (x - (((float)data->x_max +1) / 2));
-		data->map[i].y = (y - (((float)data->y_max +1) / 2));
+		data->map[i].x = (x - (((double)data->x_max +1) / 2));
+		data->map[i].y = (y - (((double)data->y_max +1) / 2));
 		data->map[i].z = ft_atoi(word[x - 1]);
 		if (data->z_min > data->map[i].z)
 			data->z_min = data->map[i].z;
